+++ src/components/StudentCourseList.vue
@@ -1,20 +1,8 @@
 <template>
     <div class="w-full max-w-4xl">
         <div class="flex">
-            <div class="w-full">
-                <div class="pt-2 mb-8 relative mx-auto text-gray-600 flex">
-                    <i class="fas fa-search absolute left-0 top-0 mt-6 ml-4"></i>
-                    <input
-                        id="message"
-                        v-model="message"
-                        class="w-full border-2 border-gray-300 bg-white h-12 px-12 rounded-lg focus:outline-none"
-                        type="search"
-                        placeholder="Filter"
-                    />
-                    <button id="refresh" class="-ml-8 my-1 focus:outline-none" title="Refresh" @click="refresh">
-                        <i class="inline fas fa-redo-alt text-gray-600 hover:text-gray-700" />
-                    </button>
-                </div>
+            <div class="w-full pt-2 mb-8">
+                <seach-bar v-model:message="message" @refresh="refresh" />
             </div>
         </div>
         <suspense>
@@ -31,6 +19,7 @@
 <script lang="ts">
     import CourseList from "./CourseList.vue";
     import LoadingComponent from "./loading/Spinner.vue";
+    import SeachBar from "./SearchBar.vue";
     import { ref } from "vue";
 
     export default {
@@ -38,6 +27,7 @@
         components: {
             CourseList,
             LoadingComponent,
+            SeachBar,
         },
         setup() {
             let message = ref("");