+++ src/components/StudentCourseList.vue
@@ -10,9 +10,7 @@
                 <courseList :key="refreshKey" />
             </template>
             <template #fallback>
-                <p class="text-center text-lg pt-20">
-                    Loading Courses...
-                </p>
+                <loading-component />
             </template>
         </suspense>
     </div>
@@ -20,6 +18,7 @@
 
 <script lang="ts">
     import CourseList from "./CourseList.vue";
+    import LoadingComponent from "./loading/Spinner.vue";
     import SeachBar from "./SearchBar.vue";
     import { ref } from "vue";
 
@@ -27,6 +26,7 @@
         name: "StudentCourseList",
         components: {
             CourseList,
+            LoadingComponent,
             SeachBar,
         },
         setup() {