+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Feature
 - add password change API + unit test (#213)
-- 
+- adds errors for nested objects, like the street of one's address (#233)
+
 # [v.0.4.4](https://github.com/upb-uc4/ui-web/compare/v0.4.3...v0.4.4) (2020-07-22)
 
 - Requires backend version 0.4.2