+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 - add password change API + unit test (#213)
 - add option for user to change the password (atm. in lecturer private profile, will be moved to a settings page) (#208)
 
+## Bug Fixes
+
+## Refactor
+- extract searchbar component for better adaptability (#131)
+
+### Usability
+- rework login error validation (#232)
 
 # [v.0.4.4](https://github.com/upb-uc4/ui-web/compare/v0.4.3...v0.4.4) (2020-07-22)
 
@@ -14,7 +21,9 @@
 ### Usability
 - change input type of semester count in student's private profile to text (#118)
 - add proper page titling based on route (#195)
+- Add a new "Add" button next to the search bar in the account/course lists for usability (#178)
 - change type of lecturer name in student's courses to button, which routes to lecturer's public profile (#115)
+
 
 ### Technical
 - declare emits in various components (adds type safety to emit)